Git_Projects/smart_parking_finder_poc/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
bcrypt==4.1.2
PyJWT==2.8.0
motor==3.5.1
email-validator==2.1.1
